ADD : "Node" - "Detach" to remove a node from a road. Works the same way as "Split" FIX : Make Osmarender work again
git-svn-id: http://svn.openstreetmap.org/applications/editors/merkaartor@12674 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4

+++ GPS/qgpssatellitetracker.cpp
@@ -137,6 +137,6 @@ void QGPSSatelliteTracker::getCoordsFromPos(int rad, int elevation, int azimuth,
     elevation = 90 - elevation;
 
     // you should know this (slept too much in trig)
-    x = cos(theta*M_PI/180) * elevation * rad / 90;
-    y = sin(theta*M_PI/180) * elevation * rad / 90;
+	x = int(cos(theta*M_PI/180) * elevation * rad / 90);
+	y = int(sin(theta*M_PI/180) * elevation * rad / 90);
 }
